🚚 Rename user hook
